requirement.txt
requests>=2.26.0,<3.0.0
beautifulsoup4>=4.9.3
selenium>=4.0.0
redis>=4.5.0
python-dotenv
aiohttp
cinemagoer
pyjsparser